[src] Rename functions of vector quantization
