+++ audyn/functional/vector_quantization.py
